pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-orchestrator"
version = "0.1.0"
description = "Human-in-the-loop coding orchestrator: OpenAI planner + Claude Code executor"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "openai>=1.30",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "anthropic>=0.28",
]

[project.scripts]
orchestrator = "orchestrator.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]
